backend/requirements.txt
flask==2.2.0
python-dotenv
flask-restful
flask-sqlalchemy==2.5.1
flask-jwt-extended
flask-mail
flask-cors
flask-login
SQLAlchemy==1.4.23
Werkzeug==2.2.2
